refactor(core): reorganize package structure and improve code modularity
패키지 구조를 재조정하여 더 나은 모듈화와 코드 이해도를 제공하도록 개선하였습니다.
기존 코드와 기능은 유지하면서도 각 컴포넌트를 분리하여 명확성을 높였습니다.

BREAKING CHANGE: 일부 클래스와 파일의 경로가 변경되었습니다. 기존 경로를 참조하거나 해당 클래스를 사용하는 경우 새로운 패키지 경로로 수정해야 합니다.
