+++ src/main/kotlin/shopping/common/CustomRestControllerAdvice.kt
@@ -1,4 +1,4 @@
-package shopping.config
+package shopping.common
 
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
