--- src/main/kotlin/shopping/exception/ProductNotFoundException.kt
@@ -1,3 +1,0 @@
-package shopping.exception
-
-class ProductNotFoundException(id: Long) : NotFoundException("product", id)
