+++ src/main/kotlin/shopping/product/dto/request/ProductRequestBody.kt
@@ -1,4 +1,4 @@
-package shopping.controller.model
+package shopping.product.dto.request
 
 import jakarta.validation.constraints.Min
 import jakarta.validation.constraints.NotBlank
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull
 import jakarta.validation.constraints.Pattern
 import jakarta.validation.constraints.Size
 import org.hibernate.validator.constraints.URL
-import shopping.repository.model.Product
+import shopping.product.entity.Product
 
 data class ProductRequestBody(
     @field:Size(
