+++ src/main/kotlin/shopping/product/exception/ProductNotFoundException.kt
@@ -1,0 +1,5 @@
+package shopping.product.exception
+
+import shopping.exception.NotFoundException
+
+class ProductNotFoundException(id: Long) : NotFoundException("product", id)
